+++ sqlx-sqlite/src/types/str.rs
@@ -94,20 +94,3 @@ impl<'q> Encode<'q, Sqlite> for Cow<'q, str> {
         Ok(IsNull::No)
     }
 }
-
-impl<'q> Encode<'q, Sqlite> for Cow<'q, [u8]> {
-    fn encode(self, args: &mut Vec<SqliteArgumentValue<'q>>) -> Result<IsNull, BoxDynError> {
-        args.push(SqliteArgumentValue::Blob(self));
-
-        Ok(IsNull::No)
-    }
-
-    fn encode_by_ref(
-        &self,
-        args: &mut Vec<SqliteArgumentValue<'q>>,
-    ) -> Result<IsNull, BoxDynError> {
-        args.push(SqliteArgumentValue::Blob(self.clone()));
-
-        Ok(IsNull::No)
-    }
-}
